+++ spec/spec_helper.rb
@@ -1,4 +1,6 @@
 require 'chefspec'
 require 'chefspec/berkshelf'
 
+CACHE = Chef::Config[:file_cache_path]
+
 at_exit { ChefSpec::Coverage.report! }