+++ spec/unit/default_spec.rb
@@ -1,91 +1,78 @@
 require 'spec_helper'
 
-SOURCE = 'https://github.com/dhoer/play-java-sample/releases/download/1.0/play-java-sample-1.0.zip'.freeze
-CONF_VARIABLES = { 'secret' => 'testingonetwothree' }.freeze
-SERVICENAME = 'sample_service'.freeze
-
-describe 'play::default' do
+describe 'play_test::default' do
   let(:chef_run) do
     ChefSpec::SoloRunner.new(
-      platform: 'centos', version: '6.7', step_into: ['play'], file_cache_path: '/var/chef/cache') do |node|
-      node.set['play']['servicename'] = SERVICENAME
-      node.set['play']['source'] = SOURCE
-      node.set['play']['conf_variables'] = CONF_VARIABLES
-    end.converge(described_recipe)
+      platform: 'centos', version: '7.0', step_into: ['play'], file_cache_path: CACHE
+    ).converge(described_recipe)
   end
-  let(:play_service_template) { chef_run.template('/etc/init.d/sample_service') }
+  let(:play_service_template) { chef_run.template('/etc/init.d/play-java-sample') }
 
-  it 'install unzip package' do
+  it 'installs play app' do
+    expect(chef_run).to install_play(
+      'https://github.com/dhoer/play-java-sample/releases/download/1.0/play-java-sample-1.0.zip'
+    ).with(
+      conf_variables:  { secret: 'mysecret' }
+    )
+  end
+
+  it 'installs unzip package' do
     expect(chef_run).to install_package('unzip')
   end
 
-  it 'install rsync package' do
-    expect(chef_run).to install_package('rsync')
+  it 'creates user' do
+    expect(chef_run).to create_user('play-java-sample')
   end
 
-  it 'creates play system user' do
-    expect(chef_run).to create_user('play')
+  it 'creates group' do
+    expect(chef_run).to create_group('play-java-sample')
   end
 
-  it 'unzips local file standalone distribution archive' do
-    expect(chef_run).to install_ark('sample_service')
+  it 'creates path' do
+    expect(chef_run).to create_directory('/opt/play/zip')
+  end
+
+  it 'downloads app' do
+    expect(chef_run).to create_remote_file("#{CACHE}/play-java-sample-1.0.zip")
+  end
+
+  it 'does not untar zip' do
+    expect(chef_run).to_not run_execute("untar #{CACHE}/play-java-sample-1.0.zip")
+  end
+
+  it 'unzips zip' do
+    expect(chef_run).to_not run_execute("unzip #{CACHE}/play-java-sample-1.0.zip")
+  end
+
+  it 'creates link' do
+    expect(chef_run).to create_link('/opt/play/zip/play-java-sample')
   end
 
   it 'verifies executable exists' do
     expect(chef_run).to run_ruby_block('verify executable exists')
   end
 
-  it 'runs play install action' do
-    expect(chef_run).to install_play(SERVICENAME).with(
-      source: SOURCE,
-      conf_variables: CONF_VARIABLES
-    )
+  it 'makes play script executable' do
+    expect(chef_run).to run_execute('make play script executable')
   end
 
-  it 'verifies conf_template exists' do
-    expect(chef_run).to run_ruby_block('verify conf_template can be run')
+  it 'configures application.conf' do
+    expect(chef_run).to create_template('/opt/play/zip/play-java-sample/conf/application.conf')
   end
 
-  describe 'service' do
-    it 'creates pid directory' do
-      expect(chef_run).to create_directory('/var/run/play').with(
-        user: 'play',
-        group: 'play'
-      )
-    end
+  it 'creates pid dir' do
+    expect(chef_run).to create_directory('/var/run/play-java-sample')
+  end
 
-    it 'does not install sample_service as a service' do
-      expect(chef_run).to_not enable_service(SERVICENAME)
-      expect(chef_run).to_not restart_service(SERVICENAME)
-      expect(chef_run).to_not stop_service(SERVICENAME)
-      expect(chef_run).to_not start_service(SERVICENAME)
-    end
+  it 'restarts service' do
+    expect(play_service_template).to notify('service[play-java-sample]').to(:restart)
+  end
 
-    it 'does not register sample_service to start on reboot' do
-      expect(chef_run).to_not run_execute('chkconfig sample_service')
-    end
+  it 'creates service script' do
+    expect(chef_run).to create_template('/etc/init.d/play-java-sample')
+  end
 
-    it 'make bin/sample_service file executable' do
-      expect(chef_run).to run_execute('chmod +x play-java-sample')
-    end
-
-    it 'creates init.d script' do
-      expect(chef_run).to create_template('/etc/init.d/sample_service').with(
-        user: 'root',
-        group: 'root',
-        backup: 5
-      )
-    end
-
-    it 'sends notifications if init.d script created' do
-      expect(play_service_template).to notify('service[sample_service]').to(:enable)
-      expect(play_service_template).to notify('service[sample_service]').to(:start)
-    end
-
-    it 'creates application.conf' do
-      expect(chef_run).to create_template('/usr/local/sample_service/conf/application.conf').with(
-        user: 'play'
-      )
-    end
+  it 'enables service script' do
+    expect(chef_run).to enable_service('play-java-sample')
   end
 end