+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-# Play Changelog
+# Changelog
+
+## 2.0.0 2016-08-15
+
+- Change resource_name default from servicename to source 
+- Default servicename to project_name when not defined 
+- Create user and group using servicename when not defined
+- Remove default recipe
+- Support systemd
+- Update upstart script
+- Update systemv script
+- Remove ark cookbook dependency
 
 ## 1.0.2 2016-04-01